Fixed bugs

+++ install.sh
@@ -29,14 +29,14 @@ sudo apt autoremove -y
 sudo apt-get install -y vim xdotool unclutter sed #install needed packages
 
 # Get JKiosk from GitHub
-section '2. JKiosk from GitHub'
+section '2. Clone JKiosk from GitHub'
 git clone https://github.com/JoelGrayson/JKiosk.git || ERR 'Could not clone git repository'
 BASE="$HOME/JKiosk"
 cd "$BASE" || ERR 'Could not install JKiosk properly'
 
 # Moves files to correct locations
 section '3. Processing JKiosk Files'
-sudo cp "$BASE/exec/system/kiosk.service" "/usr/lib/systemd/" || ERR "can't move kiosk.service"
+sudo cp "$BASE/exec/system/kiosk.service" "/usr/lib/systemd/system" || ERR "can't move kiosk.service"
 crontab "$BASE/exec/system/cronjobs" #sets cronjobs as the new crontab so turns on/off at right times and turns on kiosk on boot
 
 # Make files executable
@@ -49,7 +49,7 @@ chmod +x "$BASE/exec/relay/LOW"
 section '4. Setting Wallpaper and Splash Screen'
 # Set splash screen for when raspberry pi is booting
 SPLASH_DIR="/usr/share/plymouth/themes/pix"
-sudo mv "$SPLASH_DIR/splash.png" "$SPLASH_DIR/splash.png.bak" #backup splash saver
+[ -e "$SPLASH_DIR/splash.png" ] && sudo mv "$SPLASH_DIR/splash.png" "$SPLASH_DIR/splash.png.bak" #backup splash saver
 sudo cp "$BASE/theme/splash.png" "$SPLASH_DIR/splash.png"
 
 # Set desktop wallpaper
